requirements/local.txt
-r base.txt # inherit from base.txt
coverage==7.6.1 # measures code coverage
flake8==7.1.1 # style guide enforcement
tox==4.18.1 # test command line tool 
django-debug-toolbar==4.4.6 # display debug information
factory-boy==3.3.1 # setup test cases
tqdm==4.66.5 # progress bar for data loader script